refactor and clean code

+++ src/main/java/com/gymcrm/service/impl/TraineeServiceImpl.java
@@ -3,6 +3,7 @@ package com.gymcrm.service.impl;
 import com.gymcrm.dao.TraineeDAO;
 import com.gymcrm.model.Trainee;
 import com.gymcrm.service.TraineeService;
+import com.gymcrm.exception.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,7 @@ import java.util.Optional;
 
 @Service
 public class TraineeServiceImpl implements TraineeService {
+    private static final Logger logger = LoggerFactory.getLogger(TraineeServiceImpl.class);
     private final TraineeDAO traineeDAO;
 
     public TraineeServiceImpl(TraineeDAO traineeDAO) {
@@ -21,11 +23,16 @@ public class TraineeServiceImpl implements TraineeService {
     @Override
     public void createTrainee(Trainee trainee) {
         traineeDAO.save(trainee);
+        logger.info("Trainee with ID {} created successfully", trainee.getId());
     }
 
     @Override
     public void updateTrainee(Trainee trainee) {
+        if (!traineeDAO.findById(trainee.getId()).isPresent()) {
+            throw new ResourceNotFoundException("Trainee with ID " + trainee.getId() + " not found");
+        }
         traineeDAO.update(trainee);
+        logger.info("Trainee with ID {} updated successfully", trainee.getId());
     }
 
     @Override
@@ -40,6 +47,10 @@ public class TraineeServiceImpl implements TraineeService {
 
     @Override
     public void deleteTrainee(int id) {
+        if (!traineeDAO.findById(id).isPresent()) {
+            throw new ResourceNotFoundException("Trainee with ID " + id + " not found");
+        }
         traineeDAO.delete(id);
+        logger.info("Trainee with ID {} deleted successfully", id);
     }
 }
